requirements.txt
# Resume parsing
PyPDF2==3.0.1
docx2txt==0.8

# Gemini API
google-generativeai==0.1.0
google-api-core==2.12.0

# Google APIs (Calendar, Sheets)
google-api-python-client==2.91.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.0

# Notion integration
notion-client==2.2.1

# HTTP requests and parsing (optional for LinkedIn enrichment)
requests==2.32.0
beautifulsoup4==4.13.0

# Flask backend
Flask==2.3.3
Flask-Cors==3.0.13

# Logging and utilities
python-dotenv==1.0.0
